shopify_returns_agency/requirements.txt
flask==2.0.1
python-dotenv==0.19.0
pytest==6.2.5
agency-swarm==0.2.0
requests==2.28.1 